+++ schedule/core/src/main/kotlin/vp/togedo/data/dao/personalSchedule/FixedScheduleDao.kt
@@ -1,9 +1,9 @@
-package vp.togedo.data.dao
+package vp.togedo.data.dao.personalSchedule
 
 import org.bson.types.ObjectId
 
-data class FlexibleScheduleDao(
-    val scheduleId: ObjectId?,
+data class FixedScheduleDao(
+    var scheduleId: ObjectId?,
     val startTime: Long,
     val endTime: Long,
     val title: String,
