+++ schedule/core/src/main/kotlin/vp/togedo/service/impl/KafkaServiceImpl.kt
@@ -5,8 +5,8 @@ import org.bson.types.ObjectId
 import org.springframework.kafka.core.reactive.ReactiveKafkaProducerTemplate
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Mono
-import vp.togedo.data.dao.GroupDao
-import vp.togedo.data.dao.GroupScheduleDao
+import vp.togedo.data.dao.GroupSchedule.GroupDao
+import vp.togedo.data.dao.Group.GroupScheduleDao
 import vp.togedo.data.dto.group.InviteGroupEventDto
 import vp.togedo.data.dto.groupSchedule.CreateGroupScheduleEventDto
 import vp.togedo.service.KafkaService
