+++ schedule/presentation/src/main/kotlin/vp/togedo/connector/impl/FlexiblePersonalScheduleConnectorImpl.kt
@@ -3,7 +3,7 @@ package vp.togedo.connector.impl
 import org.bson.types.ObjectId
 import org.springframework.stereotype.Service
 import vp.togedo.connector.FlexiblePersonalScheduleConnector
-import vp.togedo.data.dao.FlexibleScheduleDao
+import vp.togedo.data.dao.personalSchedule.FlexibleScheduleDao
 import vp.togedo.data.dto.flexiblePersonalSchedule.CreateFlexibleReqDto
 import vp.togedo.data.dto.flexiblePersonalSchedule.UpdateFlexibleReqDto
 import vp.togedo.service.FlexiblePersonalScheduleService
