+++ schedule/presentation/src/main/kotlin/vp/togedo/connector/impl/GroupConnectorImpl.kt
@@ -9,7 +9,7 @@ import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import reactor.core.scheduler.Schedulers
 import vp.togedo.connector.GroupConnector
-import vp.togedo.data.dao.GroupDao
+import vp.togedo.data.dao.GroupSchedule.GroupDao
 import vp.togedo.data.dto.group.CreateGroupReqDto
 import vp.togedo.data.dto.group.GroupDto
 import vp.togedo.data.dto.group.UpdateGroupReqDto
